+++ src/main/js/pages/ChooseRestaurant.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function ChooseRestaurant() {
+    return (<>Selecione seu restaurante</>)
+}
